Move FakePIDMotor from :lib to :testing
This reduces the number of dependencies pulled in by :lib, which reduces how
much code has to be deployed to the robot.
