+++ lib/build.gradle
@@ -20,11 +20,11 @@ dependencies {
 
     testImplementation(platform('org.junit:junit-bom:5.13.1'))
     testImplementation('org.junit.jupiter:junit-jupiter')
-    testImplementation 'org.mockito:mockito-core:5.14.2'
+    testImplementation('org.mockito:mockito-core:5.14.2')
+    testImplementation('com.google.truth:truth:1.4.4')
     testRuntimeOnly('org.junit.platform:junit-platform-launcher')
     testRuntimeOnly('org.junit.vintage:junit-vintage-engine')
     testImplementation 'junit:junit:4.13.2'
-    implementation 'com.google.truth:truth:1.4.4'//needed for FakePIDMotor
     testImplementation project(':testing')
     compileOnly 'com.google.auto.value:auto-value-annotations:1.11.0'
     annotationProcessor 'com.google.auto.value:auto-value:1.11.0'
