+++ testing/src/main/java/com/team2813/lib2813/testing/FakePIDMotor.java
@@ -1,4 +1,4 @@
-package com.team2813.lib2813.util;
+package com.team2813.lib2813.testing;
 
 import com.google.common.truth.Truth;
 import com.team2813.lib2813.control.ControlMode;
